# docs/openapi.yaml
openapi: 3.0.3
info:
    title: GoChat API
    description: >
        Production-ready real-time chat API built with Go, PostgreSQL (sqlc + pgx), RabbitMQ, and WebSockets.
    version: 1.0.0
servers:
    - url: http://localhost:8080
      description: Local development server
    - url: https://api.trygochat.tech
      description: Public development server
security: []
paths:
...
    /chats:
        get:
            summary: Get User Chats
            description: Retrieves all chats (inbox) for the authenticated user, including last message preview.
            security:
                - BearerAuth: []
            responses:
                "200":
                    description: List of user chats
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ChatInboxResponse"
                "401":
                    description: Unauthorized
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "403":
                    description: Forbidden
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "500":
                    description: Internal server error
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"

        post:
            summary: Create Chat
            description: Creates a new 1-on-1 or group chat. Returns existing chat ID if 1-on-1 chat already exists between users.
            security:
                - BearerAuth: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/CreateChatRequest"
            responses:
                "201":
                    description: Chat created successfully
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Chat"
                "200":
                    description: 1-on-1 chat already exists
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    id:
                                        type: string
                                        format: uuid
                                    message:
                                        type: string
                                        example: Chat already exists
                "400":
                    description: Bad request (invalid input or 1-on-1 chat doesn't have 2 participants)
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "401":
                    description: Unauthorized
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "403":
                    description: Forbidden
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "500":
                    description: Internal server error
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
...
components:
    securitySchemes:
        BearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
            description: "Bearer token passed in the Authorization header (`Authorization: Bearer <token>`)."
        CookieAuth:
            type: apiKey
            in: cookie
            name: refresh_token
            description: "HttpOnly cookie containing the raw refresh token (`refresh_token=<token>`)."
    schemas:
...
        Chat:
            type: object
            properties:
                id:
                    type: string
                    format: uuid
                    example: 0190e241-7c9b-789a-8123-456789abcdef
                name:
                    type: string
                    nullable: true
                    example: Team Chat
                is_group:
                    type: boolean
                    example: true
                created_at:
                    type: string
                    format: date-time
        CreateChatRequest:
            type: object
            required:
                - is_group
                - participant_ids
            properties:
                name:
                    type: string
                    nullable: true
                    example: Project Alpha
                is_group:
                    type: boolean
                    example: false
                participant_ids:
                    type: array
                    items:
                        type: string
                        format: uuid
                    example:
                        - 0190e241-7c9b-789a-8123-456789abcdef
        ChatInboxItem:
            type: object
            properties:
                id:
                    type: string
                    format: uuid
                name:
                    type: string
                    nullable: true
                is_group:
                    type: boolean
                created_at:
                    type: string
                    format: date-time
                last_read_at:
                    type: string
                    format: date-time
                last_message_content:
                    type: string
                    nullable: true
                last_message_id:
                    type: string
                    format: uuid
                    nullable: true
        ChatInboxResponse:
            type: array
            items:
                $ref: "#/components/schemas/ChatInboxItem"
...
        ErrorResponse:
            type: object
            properties:
                error:
                    type: string
                    example: Error description message
...